+++ tests/commands/config/test_config_sample.py
@@ -2,11 +2,13 @@ import os
 import re
 import json
 import pytest
-import graphviz
 from unittest import mock
-from datetime import datetime
-from pathlib import Path
 import click
 
+
+# Given standard run fixtures, execute command and verify dag graph exists
+# Given invalid, existing fastq,  -> invoke cli > error
+# Given an empty reference config -> invoke command > error
+# Given analysis dir without write permissions -> invoke command > error
